line clip logic fix,clip result may be out of clip volume a little fragment attribute buffer behavior change: use negative based index,to work wih clip logic

+++ src/soft_impl/pipeline/clipper/LineClipper.hpp
@@ -37,7 +37,7 @@ namespace my_gl {
 
      	virtual ~LineClipper ();
 
-	typedef pair<float,float> ClipPercent;
+	typedef pair<double,double> ClipPercent;
 
 	static bool outOfClipVolume(const ClipPercent& clipResult);
