+++ src/soft_impl/pipeline/interpolator/WindowCoordinatesInterpolator.cpp
@@ -26,7 +26,7 @@
 namespace my_gl {
 
      WindowCoordinatesInterpolator::~WindowCoordinatesInterpolator(){}
-	  float WindowCoordinatesInterpolator::getPercent(
+	  double WindowCoordinatesInterpolator::getPercent(
 		    const CoordInfo& coord1,const CoordInfo& coord2,
 		    const LineInfo& lineInfo,
 		    const WindowCoordinates& toInterpolate)const
